+++ src/components/_modal/pages/bookmarkOrder/multi-order/Material/Material.js
@@ -13,20 +13,14 @@ import IconButton from "@/components/base/iconButton";
 import HasBeenOrderedRow from "../../../edition/hasbeenOrderedRow/HasBeenOrderedRow";
 import {
   useOrderFlow,
-  useOrderPolicy,
+  useOrderPolicyMessage,
   useOrderService,
   usePeriodica,
   usePeriodicaForm,
-  usePickupBranchId,
   useShowAlreadyOrdered,
 } from "@/components/hooks/order";
 import { useData } from "@/lib/api/api";
 import { editionManifestations } from "@/lib/api/manifestation.fragments";
-import { useBranchInfo } from "@/components/hooks/useBranchInfo";
-import { IconLink } from "@/components/base/iconlink/IconLink";
-import ExternalSvg from "@/public/icons/external_small.svg";
-import animations from "@/components/base/animation/animations.module.css";
-import { useHoldingsForAgency } from "@/components/hooks/useHoldings";
 
 /**
  * Is missing article implementation
@@ -60,14 +54,6 @@ const Material = ({
   const { service, isLoading: isLoadingOrderService } = useOrderService({
     pids,
   });
-  const { branchId } = usePickupBranchId();
-  const branch = useBranchInfo({ branchId });
-  const policy = useOrderPolicy({ branchId, pids });
-  const { branches } = useHoldingsForAgency({
-    pids,
-    agencyId: branch?.agencyId,
-  });
-  const lookupUrl = branches?.[0]?.holdings?.lookupUrl;
 
   const { data: manifestationsData, isLoading: isLoadingManifestations } =
     useData(
@@ -76,6 +62,8 @@ const Material = ({
           pid: pids,
         })
     );
+
+  const orderPolicyMessage = useOrderPolicyMessage({ pids, textType: "text4" });
 
   const isLoading = isLoadingManifestations || isLoadingAlreadyOrdered;
 
@@ -155,37 +143,12 @@ const Material = ({
     children.push(
       <>
         <Text className={styles.orderNotPossible} type="text4">
-          {policy?.physicalCopyAllowedReason === "OWNED_OWN_CATALOGUE" ? (
-            <div className={styles.path_blue}>
-              {Translate({
-                context: "localizations",
-                label: "no_pickup_allowed_for_material",
-                vars: [branch?.name],
+          {orderPolicyMessage
+            ? orderPolicyMessage
+            : Translate({
+                context: "materialcard",
+                label: "order-not-possible",
               })}
-              <IconLink
-                iconPlacement="right"
-                iconSrc={ExternalSvg}
-                iconAnimation={[
-                  animations["h-elastic"],
-                  animations["f-elastic"],
-                ]}
-                textType="text4"
-                href={lookupUrl}
-                target="_blank"
-              >
-                {Translate({
-                  context: "localizations",
-                  label: "order_locally",
-                })}
-              </IconLink>
-              .
-            </div>
-          ) : (
-            Translate({
-              context: "materialcard",
-              label: "order-not-possible",
-            })
-          )}
         </Text>
 
         <IconButton onClick={() => deleteOrder({ pids })} icon="close">
