+++ src/components/_modal/pages/bookmarkOrder/multi-order/Material/Material.module.css
@@ -3,13 +3,10 @@
   padding-top: var(--pt1);
   padding-bottom: var(--pt05);
 }
+.orderNotPossible * {
+  color: var(--error);
+}
 
 .errorLabel {
   color: var(--error_temp);
-}
-
-.path_blue,
-.path_blue * {
-  color: var(--error);
-  stroke: var(--blue);
 }
