+++ src/components/hooks/order.js
@@ -26,6 +26,12 @@ import * as orderMutations from "@/lib/api/order.mutations";
 import isEqual from "lodash/isEqual";
 import { getSessionStorageItem, setSessionStorageItem } from "@/lib/utils";
 import useDataCollect from "@/lib/useDataCollect";
+import { IconLink } from "../base/iconlink/IconLink";
+import { useHoldingsForAgency } from "./useHoldings";
+import Translate from "../base/translate/Translate";
+import ExternalSvg from "@/public/icons/external_small.svg";
+import animations from "@/components/base/animation/animations.module.css";
+import styles from "./order.module.css";
 
 /**
  * Retrieves periodica information for a list of pids
@@ -793,4 +799,47 @@ export function useOrderFlow() {
   }
 
   return { start, initialOrders, orders, setOrders, deleteOrder };
+}
+
+/**
+ * Shows a message, when order policy is false
+ */
+export function useOrderPolicyMessage({ pids, textType = "type2" }) {
+  const { branchId } = usePickupBranchId();
+  const branch = useBranchInfo({ branchId });
+  const policy = useOrderPolicy({ branchId, pids });
+  const { branches } = useHoldingsForAgency({
+    pids,
+    agencyId: branch?.agencyId,
+  });
+  const lookupUrl = branches?.[0]?.holdings?.lookupUrl;
+
+  const showMessage =
+    policy?.physicalCopyAllowedReason === "OWNED_OWN_CATALOGUE";
+
+  if (showMessage) {
+    return (
+      <div className={styles.path_blue}>
+        {Translate({
+          context: "localizations",
+          label: "no_pickup_allowed_for_material",
+          vars: [branch?.name],
+        })}
+        <IconLink
+          iconPlacement="right"
+          iconSrc={ExternalSvg}
+          iconAnimation={[animations["h-elastic"], animations["f-elastic"]]}
+          textType={textType}
+          href={lookupUrl}
+          target="_blank"
+          iconStyle={{ stroke: "blue" }}
+        >
+          {Translate({
+            context: "localizations",
+            label: "order_locally",
+          })}
+        </IconLink>
+      </div>
+    );
+  }
 }
